+++ tests/Shortcut/Unipay/QueryShortcutTest.php
@@ -2,57 +2,56 @@
 
 declare(strict_types=1);
 
-namespace Yansongda\Pay\Tests\Plugin\Unipay\Shortcut;
+namespace Yansongda\Pay\Tests\Shortcut\Unipay;
 
 use Yansongda\Pay\Exception\Exception;
 use Yansongda\Pay\Exception\InvalidParamsException;
 use Yansongda\Pay\Plugin\ParserPlugin;
 use Yansongda\Pay\Plugin\Unipay\LaunchPlugin;
+use Yansongda\Pay\Plugin\Unipay\OnlineGateway\QueryPlugin;
 use Yansongda\Pay\Plugin\Unipay\PreparePlugin;
-use Yansongda\Pay\Plugin\Unipay\QrCode\PosNormalPlugin;
-use Yansongda\Pay\Plugin\Unipay\QrCode\PosPreAuthPlugin;
 use Yansongda\Pay\Plugin\Unipay\RadarSignPlugin;
-use Yansongda\Pay\Plugin\Unipay\Shortcut\PosShortcut;
+use Yansongda\Pay\Shortcut\Unipay\QueryShortcut;
 use Yansongda\Pay\Tests\TestCase;
 
-class PosShortcutTest extends TestCase
+class QueryShortcutTest extends TestCase
 {
-    protected PosShortcut $plugin;
+    protected QueryShortcut $plugin;
 
     protected function setUp(): void
     {
         parent::setUp();
 
-        $this->plugin = new PosShortcut();
+        $this->plugin = new QueryShortcut();
     }
 
     public function testDefault()
     {
         self::assertEquals([
             PreparePlugin::class,
-            PosNormalPlugin::class,
+            QueryPlugin::class,
             RadarSignPlugin::class,
             LaunchPlugin::class,
             ParserPlugin::class,
         ], $this->plugin->getPlugins([]));
     }
 
-    public function testPreAuth()
+    public function testQrCode()
     {
         self::assertEquals([
             PreparePlugin::class,
-            PosPreAuthPlugin::class,
+            \Yansongda\Pay\Plugin\Unipay\QrCode\QueryPlugin::class,
             RadarSignPlugin::class,
             LaunchPlugin::class,
             ParserPlugin::class,
-        ], $this->plugin->getPlugins(['_action' => 'pre_auth']));
+        ], $this->plugin->getPlugins(['_action' => 'qr_code']));
     }
 
     public function testFoo()
     {
         self::expectException(InvalidParamsException::class);
         self::expectExceptionCode(Exception::PARAMS_SHORTCUT_ACTION_INVALID);
-        self::expectExceptionMessage('Pos action [fooPlugins] not supported');
+        self::expectExceptionMessage('Query action [fooPlugins] not supported');
 
         $this->plugin->getPlugins(['_action' => 'foo']);
     }