+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - change: close 的 API 参数只支持 array，不再支持 string(#901)
 - change(internal): 错误代码 const 命名规则统一(#902, #903)
 - change(internal): 微信 shortcut 从 plugin 文件夹独立出来(#904)
+- change(internal): 银联 shortcut 从 plugin 文件夹独立出来(#905)
 
 ## v3.5.3